Add files via upload
TTRD Exposed: Fake "Hacker" -who is actually a modder- mods Toontown Rewritten, a game literally made for 3 year olds, and gets mad when confronted about it. Exhibits transphobic behavior and admits to being a "pentester" in a dm with me and was openly confessing to committing a federal crime by "pentesting TTR" (Wasnt able to get a screenshot of that as he was actively trying to delete his messages to cover his tracks, proof shown in one such screenshot comparison when he deleted his transphobic comment). He was actively trying to find vulnerabilities in association with the game Toontown Rewritten, a game literally toddlers can play. Whether he was scanning their website for vulnerabilities or other servers/networks in the system is beyond me. He was doing it without express permission from the toontown rewritten site's operators and openly said this to me in a DM and that he even presented his findings but that the TTR team was "very hostile" to him for doing so. As such, if this is true, he has openly confessed to literally breaking the law on cybercrime. Screenshots of his "hack/exploit" involves a photo of an .MP4 file in which I witnessed the mod take place. He has since deleted that .MP4 file and I was not able to get a copy in time to leak.